% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aics.R
\docType{package}
\encoding{UTF-8}
\name{aics-package}
\alias{aics-package}
\alias{aics}
\title{aics: Ancestry Inference from Cancer Sequencing}
\description{
The aics package implements specialized algorithms that enable
ancestry inference from various cancer sequencing sources (RNA,
Exome and Whole-Genome sequencing).
}
\details{
The aics package also implements simulation algorithm that generates
synthetic cancer-derived data.
}
\seealso{
\itemize{
    \item \code{\link{TODO}} {for TODO}
}
}
\author{
Pascal Belleau, Astrid Deschênes and
Alexander Krasnitz

Maintainer:
Pascal Belleau <pascal_belleau@hotmail.com>
}
\keyword{package}
